добавил абстрактный класс и добавил наследование простым компонентам
