+++ src/view/button-show-more-view.js
@@ -1,25 +1,11 @@
-import { createElement } from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createButtonShowMoreTemplate = () => (
   '<button class="films-list__show-more">Show more</button>'
 );
 
-export default class ButtonShowMoreView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class ButtonShowMoreView extends AbstractView {
   get template() {
     return createButtonShowMoreTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
