+++ src/view/films-list-container.js
@@ -1,23 +1,9 @@
-import { createElement } from '../render.js';
+import AbstractView from './abstract-view';
 
 const createFilmsListContainerTemplate = () => '<div class="films-list__container"></div>';
 
-export default class FilmsListContainerView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class FilmsListContainerView extends AbstractView {
   get template() {
     return createFilmsListContainerTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
