+++ src/view/films-list-title-view.js
@@ -1,24 +1,10 @@
-import { createElement } from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createFilmsListTitleTemplate = () => '<h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>';
 
-export default class FilmsListTitleView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class FilmsListTitleView extends AbstractView {
   get template() {
     return createFilmsListTitleTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
