+++ src/view/films-list-view.js
@@ -1,23 +1,9 @@
-import { createElement } from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createFilmsListTemplate = () => '<section class="films-list"></section>';
 
-export default class FilmsListView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class FilmsListView extends AbstractView {
   get template() {
     return createFilmsListTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
