+++ src/view/films-view.js
@@ -1,23 +1,9 @@
-import { createElement } from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createFilmsTemplate = () => '<section class="films"></section>';
 
-export default class FilmsView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class FilmsView extends AbstractView {
   get template() {
     return createFilmsTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
