+++ src/view/no-films-view.js
@@ -1,23 +1,9 @@
-import { createElement } from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const noFilmsTemplate = () => '<h2 class="films-list__title">There are no movies in our database</h2>';
 
-export default class NoFilmsView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class NoFilmsView extends AbstractView {
   get template() {
     return noFilmsTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
