+++ src/view/profile-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createProfileTemplate = () => (
   `<section class="header__profile profile">
@@ -7,23 +7,9 @@ const createProfileTemplate = () => (
 </section>`
 );
 
-export default class ProfileView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class ProfileView extends AbstractView {
   get template() {
     return createProfileTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
